% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flxbetabinom.R
\name{FLXMCbetabinom}
\alias{FLXMCbetabinom}
\title{FlexMix driver for regularized beta-binomial mixtures}
\usage{
FLXMCbetabinom(
  formula = . ~ .,
  size,
  alpha2 = 0,
  eps = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{size}{number of trials (zero or more)}

\item{alpha2}{Regularization parameter. Can be regarded the same as
adding \code{alpha2} observations conforming to the population mean to each
component.}

\item{eps}{Lower threshold for the shape parameters a and b}
}
\value{
an object of class FLXC
}
\description{
This model driver can be used to cluster data using the beta-binomial
distribution.
}
\details{
Using a regularization parameter \code{alpha2} greater than zero
acts as adding \code{alpha2} observations conforming to the population
mean to each component. This can be used to avoid degenerate
solutions. It also has the effect
that clusters become more similar to each other the larger
\code{alpha2} is chosen. For small values it is mostly negligible however.
}
\references{
\itemize{
\item Ernst, D, Ortega Menjivar, L, Scharl T, Grün, B (2025).
\emph{Ordinal clustering with the flex-Scheme.}
Austrian Statistics Journal. \emph{Submitted manuscript}.
\item Kondofersky, Ivan (2008).
\emph{Modellbasiertes Clustern mit der Beta-Binomialverteilung.}
Bachelor's thesis, Ludwig-Maximilians-Universität München
}
}
