% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flxbinomial.R
\name{FLXMCbinomial}
\alias{FLXMCbinomial}
\title{FlexMix driver for regularized binomial mixtures}
\usage{
FLXMCbinomial(formula = . ~ ., size = NULL, alpha2 = 0, eps = 0)
}
\arguments{
\item{size}{number of trials (zero or more)}

\item{alpha2}{Regularization parameter. Can be regarded the same as
adding \code{alpha2} observations conforming to the population mean to each
component.}

\item{eps}{When greater than zero, probabilities are truncated to fit
in [eps, 1-eps]}
}
\value{
an object of class FLXC
}
\description{
This model driver can be used to cluster binomial data.
Using a regularization parameter \code{alpha2} greater than zero
acts as adding \code{alpha2} observations conforming to the population
mean to each component. This can be used to avoid degenerate
solutions (i.e. probabilites of 0 or 1). It also has the effect
that clusters become more similar to each other the larger
\code{alpha2} is chosen. For small values it is mostly negligible however.
}
\details{
Parameter estimation is achieved using the MAP estimator for each cluster
and variable using a Beta prior.
}
\references{
\itemize{
\item Ernst, D, Ortega Menjivar, L, Scharl T, Grün, B (2025).
\emph{Ordinal clustering with the flex-Scheme.}
Austrian Statistics Journal. \emph{Submitted manuscript}.
}
}
