% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flxregnorm.R
\name{FLXMCregnorm_defaults}
\alias{FLXMCregnorm_defaults}
\title{Default parameters for regularized normal mixtures based on the data set}
\usage{
FLXMCregnorm_defaults(x, zeta_p = NULL, kappa_p = 0.01, nu_p = 3, G = NULL)
}
\arguments{
\item{x}{The data set to be clustered. Should be the same data set as is
used in flexmix's model formula.}

\item{zeta_p}{Scale (hyperparameter for IG prior). If not given the empirical
variance divided by the square of the number of components
is used as per Fraley and Raftery (2007).}

\item{kappa_p}{Shrinkage parameter. Functions as if you added
\code{kappa_p} observations according to the population mean to
each component (hyperparameter for IG prior)}

\item{nu_p}{Degress of freedom (hyperparameter for IG prior)}

\item{G}{Number of components in the mixture model (not used if zeta_p is given)}
}
\description{
Default parameters for regularized normal mixtures based on the data set
}
