% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kccaExtendedFamily.R
\name{kccaExtendedFamily}
\alias{kccaExtendedFamily}
\title{Extending K-Centroids clustering to (mixed-with-)ordinal data}
\usage{
kccaExtendedFamily(which=c('kModes', 'kGDM2', 'kGower'),
                   cent=NULL,
                   preproc=NULL,
                   xrange=NULL,
                   xmethods=NULL,
                   trim=0, groupFun='minSumClusters')
flexclust::kcca(x, k,
                family=kccaExtendedFamily(which=c('kModes', 'kGDM2', 'kGower'), ...))
}
\arguments{
\item{which}{One of either \code{kModes}, \code{kGDM2} or \code{kGower}, the three predefined
methods for K-centroids clustering.
For more information on each of them, see the Details section.}

\item{cent}{Function for determining cluster
centroids (See: \href{https://search.r-project.org/CRAN/refmans/flexclust/html/kcca.html}{flexclust::kccaFamily()}).
This parameter is ignored for \code{which='kModes'}, and \code{centMode} is used.
For \code{kGDM2} and \code{kGower}, \code{cent=NULL} defaults to a general purpose optimizer.}

\item{preproc}{Preprocessing function to be applied to the data before clustering
(See: \href{https://search.r-project.org/CRAN/refmans/flexclust/html/kcca.html}{flexclust::kccaFamily()}).
This parameter is ignored for \code{which='kGower'}, instead, the default
preprocessing proposed by Gower (1971) and Kaufman\&Rousseeuw (1990)
is conducted. For \code{kGDM2} and \code{kModes}, users can specify preprocessing
steps here, though this is not recommended.}

\item{xrange}{The range of the data in \code{x}. Options are:
- 'all', which will use the whole range of the data object \code{x},
- 'columnwise', which will use columnwise ranges of the data object \code{x},
- a vector of \code{c(min,max)}, which allows the user to specify the
absolute possible minimum and maximum of the data object \code{x},
- or a list of vectors \code{list(c(min,max), c(min,max),...)} with length
\code{ncol(x)}, where columnwise ranges can be specified by the user.
This parameter takes effect for \code{which='kGDM2'} and \code{which='kGower'},
and has no effect for \code{which='kModes'}.
\code{xrange=NULL} defaults to \code{all} for \code{kGDM2}, and to \code{columnwise} for \code{kGower}.}

\item{xmethods}{An optional character vector of length \code{ncol(x)} that specifies
the distance measure for each column of \code{x}. Currently only used for
\code{kGower}. For \code{kGower}, \code{xmethods=NULL} results in the use of
default methods for each column of \code{x}. For more information on
allowed input values, and default measures, see the Details section.}

\item{trim}{See \code{\link[flexclust:kcca]{flexclust::kccaFamily()}}: Proportion of points trimmed in
robust clustering.}

\item{groupFun}{See \code{\link[flexclust:kcca]{flexclust::kccaFamily()}}: A character string specifying
the function for clustering. Default is \code{'minSumClusters'}.}
}
\value{
An object of class \code{flexclust::kccaFamily}. When using it within \code{flexclust::kcca},
\code{kccaExtendedFamily(which='kModes')} will result in an object able to conduct
kModes clustering, \code{which='kGDM2'} will result in an object for K-centroids clustering using
GDM2 distance, and \code{which='kGower'} will result in an object for k-centroids clustering using
Gower's distance.
The output of \code{flexclust::kcca} will be an object of class \code{kcca}, and can thus
be used in \code{flexclust}'s plotting, stepwise-clustering, and bootstrapping methods.
}
\description{
This wrapper creates objects of class \code{flexclust::kccaFamily},
which can be used within \code{flexclust::kcca} to conduct K-centroids
clustering using the following methods:
\itemize{
\item \strong{kModes} (after Weihs et\~al., 2005)
\item \strong{kGower} (Gower's distance after Kaufman \& Rousseeuw, 1990,
and a user specified centroid)
\item \strong{kGDM2} (GDM2 distance after Walesiak et\~al., 1993, and a
user specified centroid)
}
}
\details{
\code{kccaExtendedFamily(which='kModes')} creates an object that will perform
\strong{kModes} clustering, that is to say K-centroids clustering using Simple
Matching Distance (counts of disagreements) and modes as centroids.
User-specified alternative centroids in parameter \code{cent} are ignored for
this method.

An object created with \verb{which='kGower} will use Gower's method as described
in Kaufman\&Rousseeuw (1990) for clustering:
\itemize{
\item Numeric and/or ordinal variables are scaled by \eqn{\frac{\mathbf{x}-\min{\mathbf{x}}}{\max{\mathbf{x}-\min{\mathbf{x}}}}}
(please note that ordinal variables have to be coded so that they represent their ranks
in the way that they start with 1 and go in steps of one up to their maximum level \eqn{m}.)
\item Distances are calculated for each column \eqn{p} (squared Euclidean distance, \code{distEuclidean},
is recommended for numeric, Manhattan distance, \code{distManhattan} for ordinal, Simple Matching
Distance, \code{distSimMatch} for categorical, and Jaccard distance, \code{distJaccard} for asymmetric
binary variables), and they are summed up as:
\deqn{d(x_i, x_k) = \frac{\sum_{j=1}^p \delta_{ikj} d(x_{ij}, x_{kj})}{\sum_{j=1}^p \delta_{ikj}}}
with the weight \eqn{\delta_{ikj}} being 1 if both values \eqn{x_{ij}} and \eqn{x_{kj}} are
not missing, and in the case of asymmetric binary variables, at least one of them is not 0.

The columnwise distances used can be influenced twofold: By passing a character
vector of length \eqn{p} to \code{xmethods} that specifies the distance for each column.
Options are: \code{distEuclidean}, \code{distManhattan}, \code{distJaccard}, and \code{distSimMatch}.
Another option is to not specify any methods within \code{kccaExtendedFamily}, but rather
in \code{kcca} pass a dataframe to \code{x}, where each column is coded to achieve the desired
treatment. \code{distEuclidean} is used on numeric and integer columns, \code{distManhattan}
on columns that are coded as ordered factors, \code{distSimMatch} is the default for categorically
coded columns, and \code{distJaccard} is the default for binary coded columns.

For this method, if \code{cent=NULL}, a general purpose optimizer with \code{NA} omission
will be applied for centroid calculation.
}

An object created with \code{which=kGDM2} will use the GDM2 distance for ordinal variables,
which was first introduced by Walesiak et\~al. (1993), and adapted in Ernst et\~al. (2025),
as the distance measure within \code{flexclust::kcca}.

The principle behind it is that the ordinality of a variable will be
respected by conducting only relational operations on them, such as \eqn{\leq}, \eqn{\geq} and \eqn{=}.
By translating \eqn{x} to its relative frequencies and empirical cumulative
distributions, we are able to extend this principle to compare two arbitrary
values, and thus conduct K-Centroids clustering. For more details, see Ernst et\~al. (2025).

Also for this method, if \code{cent=NULL}, a general purpose optimizer with \code{NA} omission
will be applied for centroid calculation.

\strong{Scale handling}
In \code{kModes}, all variables are treated as unordered factors.
In \code{kGDM2}, all variables are treated as ordered factors, with strict assumptions
regarding their ordinality.
\code{kGower} is currently the only method designed to handle mixed-type data. For ordinal
variables, the assumptions are more lax than with GDM2 distance.

\strong{NA handling}
NA handling via omission and upweighting non-missing variables is currently
only implemented for \code{kGower}. Within \code{kModes}, the omission of NA responses
can be avoided by coding missings as valid factor levels. For \code{kGDM2}, currently
the only option is to omit missing values completely.
}
\examples{
# Example 1: kModes
set.seed(123)
dat <- data.frame(cont = sample(1:100, 10, replace=TRUE)/10,
                  bin_sym = as.logical(sample(0:1, 10, replace=TRUE)),
                  bin_asym = as.logical(sample(0:1, 10, replace=TRUE)),                     
                  ord_levmis = factor(sample(1:5, 10, replace=TRUE),
                                      levels=1:6, ordered=TRUE),
                  ord_levfull = factor(sample(1:4, 10, replace=TRUE),
                                       levels=1:4, ordered=TRUE),
                  nom = factor(sample(letters[1:4], 10, replace=TRUE),
                               levels=letters[1:4]))
flexclust::kcca(dat, k=3, family=kccaExtendedFamily('kModes'))

# Example 2: kGDM2
flexclust::kcca(dat, k=3, family=kccaExtendedFamily('kGDM2'))
flexclust::kcca(dat, k=3, family=kccaExtendedFamily('kGDM2',
                                                    xrange='columnwise'))
# Example 3: kGower
flexclust::kcca(dat, 3, kccaExtendedFamily('kGower'))
nas <- sample(c(TRUE,FALSE), prod(dim(dat)), replace=TRUE, prob=c(0.1,0.9)) |> 
   matrix(nrow=nrow(dat))
dat[nas] <- NA
flexclust::kcca(dat, 3, kccaExtendedFamily('kGower'))
flexclust::kcca(dat, 3, kccaExtendedFamily('kGower',
                                           xrange='all'))
flexclust::kcca(dat, 3, kccaExtendedFamily('kGower',
                                           xmethods=c('distEuclidean',
                                                      'distEuclidean',
                                                      'distJaccard',
                                                       'distManhattan',
                                                       'distManhattan',
                                                       'distSimMatch')))
#the case where column 2 is a binary variable, but is symmetric

}
\references{
\itemize{
\item Ernst, D, Ortega Menjivar, L, Scharl T, Grün, B (2025).
\emph{Ordinal clustering with the flex-Scheme.}
Austrian Statistics Journal. \emph{Submitted manuscript}.
\item Gower, JC (1971).
\emph{A General Coefficient for Similarity and Some of Its Properties.}
Biometrics, 27(4), 857-871.
\doi{doi:10.2307/2528823}
\item Kaufman, L, Rousseeuw, P (1990).
\emph{Finding Groups in Data: An Introduction to Cluster Analysis.}
Wiley Series in Probability and Statistics.
\doi{doi:10.1002/9780470316801}
\item Leisch, F (2006). \emph{A Toolbox for K-Centroids Cluster Analysis.}
Computational Statistics and Data Analysis, 17(3), 526-544.
\doi{doi:10.1016/j.csda.2005.10.006}
\item Walesiak, M (1993). \emph{Statystyczna Analiza Wielowymiarowa w Badaniach Marketingowych.}
Wydawnictwo Akademii Ekonomicznej, 44-46.
\item Weihs, C, Ligges, U, Luebke, K, Raabe, N (2005).
\emph{klaR Analyzing German Business Cycles.} In: Data Analysis an
Decision Support, Springer: Berlin. 335-343.
\doi{doi:10.1007/3-540-28397-8_36}
}
}
\seealso{
\href{https://search.r-project.org/CRAN/refmans/flexclust/html/kcca.html}{flexclust::kcca()}
\href{https://search.r-project.org/CRAN/refmans/flexclust/html/stepFlexclust.html}{flexclust::stepFlexclust()}
\href{https://search.r-project.org/CRAN/refmans/flexclust/html/bootFlexclust.html}{flexclust::bootFlexclust()}
}
