% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroidFunctions.R
\name{centMode}
\alias{centMode}
\alias{centMin}
\alias{centOptimNA}
\alias{centroids}
\title{Centroid Functions for K-Centroids Clustering of (Ordinal) Categorical/Mixed Data}
\usage{
centMode(x)

centMin(x, dist, xrange = NULL)

centOptimNA(x, dist)
}
\arguments{
\item{x}{A numeric matrix or data frame. Categorical/ordinal variables
need to be coded as \code{1:length(levels(x[,i]))} in steps of one.}

\item{dist}{The distance measure function to be minimized in \code{centMin}.}

\item{xrange}{Range specification for the variables. Currently only used for \code{centMin}.
Possible values are:
\itemize{
\item \code{NULL} (default): defaults to \code{'all'}.
\item \code{'all'}: Uses the range of the entire data set.
\item \code{'columnwise'}: Uses column-specific ranges.
\item A numeric vector of c(min, max): Applies the specified range to all columns.
\item A list of numeric vectors of c(min,max): Uses the user-specified columnwise
ranges, the length of the list must be equal to the number of columns to be scaled.
}}
}
\value{
A named numeric vector containing the centroid values for each column of \code{x}.
}
\description{
Functions to calculate cluster centroids for K-centroids clustering that extend the
options available in package \pkg{flexclust}.

\code{centMode} calculates centroids based on the mode of each variable.
\code{centMin} minimizes the applied distance metric to find centroids within a specified range.
\code{centOptimNA} replicates the exact behaviour of \code{\link[flexclust:distances]{flexclust::centOptim()}}, just with NA removal.

These functions are designed for use within \code{\link[flexclust:kcca]{flexclust::kcca()}} or functions that are built
upon it. Their use is easiest via the wrapper \code{kccaExtendedFamily}, for more information
see \code{\link[=kccaExtendedFamily]{kccaExtendedFamily()}}.
}
\details{
\itemize{
\item \strong{\code{centMode}}: Column-wise modes are used as centroids, and ties are
broken randomly. In combination with Simple Matching Distance (\code{distSimMatch}),
this results in the \code{kmodes} algorithm.
\item \strong{\code{centMin}}: Column-wise centroids are calculated by minimizing the
specified distance measure between a value of \code{x}, and all possible levels of \code{x}.
\item \strong{centOptimNA}: Column-wise centroids are calculated by minimizing the
specified distance measure via a general purpose optimizer. Unlike in \code{\link[flexclust:distances]{flexclust::centOptim()}},
NAs are removed from the starting search values.
}
}
\examples{
# Example: Mode as centroid
dat <- data.frame(A = rep(2:5, 2),
                  B = rep(1:4, 2),
                  C = rep(c(1, 2, 4, 5), 2))
centMode(dat)
## within kcca
flexclust::kcca(dat, 3, family=kccaExtendedFamily('kModes')) #default centroid

# Example: Centroid is level for which distance is minimal
centMin(dat, flexclust::distManhattan, xrange = 'all')
## within kcca
flexclust::kcca(dat, 3,
                family=flexclust::kccaFamily(dist=flexclust::distManhattan,
                                             cent=\(y) centMin(y, flexclust::distManhattan,
                                                               xrange='all')))
                             
# Example: Centroid calculated by general purpose optimizer with NA removal
nas <- sample(c(TRUE, FALSE), prod(dim(dat)),
              replace=TRUE, prob=c(0.1,0.9)) |> 
       matrix(nrow=nrow(dat))
dat[nas] <- NA
centOptimNA(dat, flexclust::distManhattan)
## within kcca
flexclust::kcca(dat, 3, family=kccaExtendedFamily('kGower')) #default centroid
}
\seealso{
\code{\link[=kccaExtendedFamily]{kccaExtendedFamily()}},
\code{\link[flexclust:kcca]{flexclust::kcca()}}
}
