% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flxregnorm.R
\name{FLXMCregnorm}
\alias{FLXMCregnorm}
\title{FlexMix driver for regularized multivariate normal mixtures}
\usage{
FLXMCregnorm(formula = . ~ ., xi_p = NULL, kappa_p = 0.01, nu_p = 3, G = NULL)
}
\arguments{
\item{formula}{A formula describing the normal components}

\item{kappa_p}{Regularization parameter. Functions as if you added
kappa_p observations according to the population mean to the
data}

\item{nu_p}{Regularization parameter.}

\item{G}{Number of components in the mixture model (not used if xi_p is given)}
}
\value{
an object of class FLXC
}
\description{
This model implements the regularization method as introduced in
Fraley & Raftery (...) for multivariate normal mixtures.
The covariance matrix for each component is assumed to be diagonal.
}
\references{
\itemize{
\item Ernst, D, Ortega Menjivar, L, Scharl T, Grün, B (2025).
\emph{Ordinal clustering with the flex-Scheme.}
Austrian Statistics Journal. \emph{Submitted manuscript}.
\item Fraley, C, Raftery AE (2007)
\emph{Bayesian Regularization for Normal Mixture Estimation and Model-Based Clustering.}
Journal of Classification, 24(2), 155-181
}
}
